Refactor Contains function to improve performance by using index iteration over collection