@lennon-guan 2024-02-19
// Benchmark code & result