@samber 2024-06-27
Wow. I didn't expect a 30% speed-up!!

Thanks for the tips!