@andig 2024-06-29
With `SliceToChannel` I'm really wondering what's being sent down to the channel. If that is not an allocated copy, then the semantics of the functions should have changed by this PR. If it is, I'm wondering where the speedup comes from?

With `IndexOf` I'm wondering why the compiler shouldn't recognize that the comparison is a read-only operation and just not do the allocation?

Every time I think I understand go I find out not...