@andig 2024-06-29
Since I was curious I ran the contains benchmarks on Darwin:

```
goos: darwin
goarch: arm64
                 │   old.txt   │               new.txt                │
                 │   sec/op    │    sec/op     vs base                │
ContainsInt-8      5.370n ± 6%   5.325n ±  0%   -0.82% (p=0.045 n=10)
ContainsString-8   7.320n ± 5%   7.224n ±  5%        ~ (p=0.210 n=10)
ContainsStruct-8   26.93n ± 1%   19.00n ± 11%  -29.44% (p=0.000 n=10)
geomean            10.19n        9.009n        -11.61%

```